+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Disable API validation for snapshots
 
 ### Changed
 - `publication.platform` changed to `publishing.platform`